% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixconstruction.R
\name{matrix_product_lit}
\alias{matrix_product_lit}
\title{Computing the product of two matrices with symbolic coefficients}
\usage{
matrix_product_lit(A, B)
}
\arguments{
\item{A}{Matrix of character strings}

\item{B}{Matrix of character strings}
}
\value{
The product AB cleaned of product with "0" and "1"
}
\description{
This function computes the product of two squares matrices of same size with character strings as coefficients and simplifies product with "0" or "1"
}
